requirements.txt
aiortc>=1.9.0
aiohttp>=3.9.0
pycryptodome>=3.20.0
Pillow>=11.0.0
numpy
